[performance-reports] switching to printf

+++ test_src/performance_test.h
@@ -33,7 +33,7 @@ void startTimer(uint8_t granularity);
 uint64_t stopTimer(void);
 void getOverhead(uint16_t *constoh, uint16_t *intoh);
 
-void print_time_P(PGM_P s, uint64_t t);
+void print_time_P(PGM_P s, uint32_t t);
 void print_overhead(void);
 
 #endif /*PERFORMANCE_TEST_H_*/
